refactor: split oversized Wahoo webhook module
